+++ packages/eslint-plugin/lib/configs/rules/stylistic-issues.js
@@ -36,6 +36,10 @@ module.exports = {
 		}
 	} ],
 	
+	// Enforce consistent indentation
+	// https://eslint.org/docs/rules/indent
+	indent: [ 'error', 'tab', { SwitchCase: 1 } ],
+	
 	// Specify whether double or single quotes should be used in JSX attributes
 	// https://eslint.org/docs/rules/jsx-quotes
 	'jsx-quotes': [ 'error', 'prefer-double' ],