+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 - Added `comma-spacing` to enforce consistent spacing before and after commas
+- Added `indent` to enforce consistent indentation
 
 # 0.0.4
 - Added `comma-dangle` to require trailing commas